bring relational compare app back to life
it generates the same output as before, but based on the json output, and not
on the internal report method.

had to add some new info into the json, which is not yet on the schema. I need to figure
out how to modify the schemas.

tested with spring target 4 which adds a trampoline, and aaron's target 4 binary for
novemeber hackathon which adds two instructions so the cfg is still the same. Both
produced the expected output. Have not tested it with a binary that has more than
one function changed