make the assembly instructions available in JSONInstructionComparison objects
this also makes instr2 optional which is the case for instructions that
were deleted. I'm not sure if instr1 should also be optional for
instructions that were added? I don't have a test case for that.

related to that, this also makes InstructionRelationalAnalysis be
aware that there may not be a second instruction before blindly
accessing it.