A massive change to try to unify all json output
I went a bit cray cray and decided that the format we had for CFG
comparison in json provided a good platform to unify all json output for
the relational analysis. So this giant change does that. It changes the
json functions in relational used by compare app and compare functions
to use the format that is currently produced in jsonrelutil, and add
their own information. Then it changes the jsoninterface code to read in
that json and produce a text report that's as close to the existing one
as I could make it.

In addition to the json format changes, this teaches the report visitor
to go all the way into the instruction comparison and assembly level.
that part is not the cleanest, but the visitor approach makes it hard to
generate two separate sets of output as we have to do here: first the
block-level stats, then the instruction-level information.

This also does a lot of changes to the BlockRelationalAnalysis code.
There was conflict between what the JSON side expects to be generated
and what is actually being generated. I changed the JSON side to match,
hopefully that was the right choice. I also changed the data being
generated so we can fill in all the information required by the final
report, in particular instructions added and removed.

Another change is that since we use the same report object to generate
the app comparison and function comparison, I had to split off the
'details' flag that we had before to give us more control:
- show_blocks: show stats at the block level. this produces the default
report for 'compare function'
- 'show_instrs': show stats at the instruction level. this corresponds
to passing '--details' to 'compare function'
- if both flags are off, we get the report for 'relational compare app'

Finally, I piped in the callees functionality that was present before
in report() so that the json output reflects that. It's untested tho
as I don't have an example to try it out on.