switch cfg comparison's json path to use the unified json file
instead of the custom json function from jsonresultutil. note that some of the
functions in there are used by the non-relational case so those cannot be deleted.

also, the dot/pdf path does not use the json output to do its job yet. I'll
leave that to be tackled later on